requirements.txt
ehforwarderbot
itchat
python-magic
pillow
pyqrcode
xmltodict
PyYaml